adding chat fixes, refactoring content analytics screens for better reusability and scability
