+++ app/(dashboard)/_components/content-analytics/index.tsx
@@ -1,0 +1,1 @@
+export * from './screens/ContentAnalyticsScreen';
